@pacoxu 2025-07-22
/cc @aojea @SergeyKanzhelev 
per your comments https://github.com/kubernetes/kubernetes/issues/115939#issuecomment-3094823643