@Ahmed-Elgamel 2025-08-21
I added a benchmark function cad75c477730e676c7f089233feae23c8df6f7ff to evaluate this approach for optimizing HTTP probes. The benchmark follows the same structure from previous bencmharks for this issue.

```
goos: darwin
goarch: arm64
pkg: k8s.io/kubernetes/pkg/kubelet/prober
cpu: Apple M2
                        │ nocaching.txt │    caching.txt     │
                        │    sec/op     │   sec/op     vs base   │
WorkerProbe_NoCaching-8     21.17m ± 3%
WorkerProbe_Caching-8                     250.8µ ± 2%
geomean                     21.17m        250.8µ       ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

                        │ nocaching.txt │     caching.txt     │
                        │     B/op      │     B/op      vs base   │
WorkerProbe_NoCaching-8    74.69Mi ± 0%
WorkerProbe_Caching-8                     385.8Ki ± 0%
geomean                    74.69Mi        385.8Ki       ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

                        │ nocaching.txt │    caching.txt     │
                        │   allocs/op   │  allocs/op   vs base   │
WorkerProbe_NoCaching-8     859.1k ± 0%
WorkerProbe_Caching-8                     3.634k ± 0%
geomean                     859.1k        3.634k       ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean
```